Move ANGLE name shenanigans to WebGL thread
